update: webMain, UI; refactor: clean build.gradle.kts, libs.versions.toml

+++ src/composeApp/src/commonMain/kotlin/hcmus/bugscanner/data/repository/EncyclopediaRepositoryImpl.kt
@@ -12,7 +12,7 @@ class EncyclopediaRepositoryImpl : EncyclopediaRepository {
     private val db = Firebase.firestore
     private val encyclopediaCollection = db.collection("encyclopedia")
 
-    override suspend fun getExploreInsects(searchQuery: String, limit: Long): List<BugInfo> {
+    override suspend fun getExploreInsects(searchQuery: String, limit: Int): List<BugInfo> {
         return try {
             val query = if (searchQuery.isNotBlank()) {
                 val searchStr = searchQuery.trim()
