+++ src/composeApp/src/commonMain/kotlin/hcmus/bugscanner/domain/repository/EncyclopediaRepository.kt
@@ -6,7 +6,7 @@ import hcmus.bugscanner.domain.model.BugInfo
  * Interface định nghĩa các thao tác với Bách khoa toàn thư
  */
 interface EncyclopediaRepository {
-    suspend fun getExploreInsects(searchQuery: String = "", limit: Long = 20): List<BugInfo>
+    suspend fun getExploreInsects(searchQuery: String = "", limit: Int = 20): List<BugInfo>
     suspend fun getBugByName(name: String): BugInfo?
     suspend fun getBugByScientificName(scientificName: String): BugInfo?
 }
